+++ includes/Service/Export/ExportService.php
@@ -14,9 +14,9 @@
 
 namespace SilverAssist\ContactFormToAPI\Service\Export;
 
-use SilverAssist\ContactFormToAPI\Core\Interfaces\LoadableInterface;
 use SilverAssist\ContactFormToAPI\Service\Logging\LogReader;
 use SilverAssist\ContactFormToAPI\Service\Security\SensitiveDataPatterns;
+use SilverAssist\PluginKernel\Interfaces\LoadableInterface;
 
 \defined( 'ABSPATH' ) || exit;